M1D14 - lecture code and slides.

+++ module-1/14_Managing_Inheritance/lecture/src/main/java/com/techelevator/farm/Sellable.java
@@ -1,8 +1,6 @@
 package com.techelevator.farm;
 
-import java.math.BigDecimal;
-
 public interface Sellable {
     String getName();
-    BigDecimal getPrice();
+    int getPrice();
 }
